frontend/requirements.txt
streamlit>=1.35.0
requests
python-dotenv
sseclient-py